problem-solver/py/requirements.txt
py-sc-kpm==0.2.0
py-sc-client==0.3.0
requests
numpy
tensorflow
tf-keras
pandas
langchain
langchain-core
langchain_community
langchain_chroma
langchain-huggingface
docx2txt
pypdf
python-multipart
sentence-transformers
fastapi
uvicorn

pdfminer.six
matplotlib
